Created a 2nd version of BernsteinPolynomial that receives its storage as argument.
This is to allow the caller to allocate once and use for many calls.
This is expected to improve performance considerably.

There is still a version that does not take the extra argument. That version
allocates memory internally and calls the other to do the job.

--HG--
branch : BernsteinPolynomial memory allocation